Fix Hookをファイルに切り出し
